+++ airflow/kubernetes/pod_generator_deprecated.py
@@ -25,11 +25,12 @@ is supported and no serialization need be written.
 from __future__ import annotations
 
 import copy
-import hashlib
 import re
 import uuid
 
 from kubernetes.client import models as k8s
+
+from airflow.utils.hashlib_wrapper import md5
 
 MAX_POD_ID_LEN = 253
 
@@ -72,7 +73,7 @@ def make_safe_label_value(string):
     safe_label = re.sub(r"^[^a-z0-9A-Z]*|[^a-zA-Z0-9_\-\.]|[^a-z0-9A-Z]*$", "", string)
 
     if len(safe_label) > MAX_LABEL_LEN or string != safe_label:
-        safe_hash = hashlib.md5(string.encode()).hexdigest()[:9]
+        safe_hash = md5(string.encode()).hexdigest()[:9]
         safe_label = safe_label[: MAX_LABEL_LEN - len(safe_hash) - 1] + "-" + safe_hash
 
     return safe_label