+++ airflow/www/static/js/dag/details/index.tsx
@@ -33,7 +33,7 @@ import { useSearchParams } from "react-router-dom";
 import useSelection from "src/dag/useSelection";
 import { getTask, getMetaValue } from "src/utils";
 import { useGridData, useTaskInstance } from "src/api";
-import { MdDetails, MdAccountTree, MdReorder } from "react-icons/md";
+import { MdDetails, MdAccountTree, MdReorder, MdCode } from "react-icons/md";
 import { BiBracket } from "react-icons/bi";
 import URLSearchParamsWrapper from "src/utils/URLSearchParamWrapper";
 
@@ -42,6 +42,7 @@ import TaskInstanceContent from "./taskInstance";
 import DagRunContent from "./dagRun";
 import DagContent from "./Dag";
 import Graph from "./graph";
+import DagCode from "./dagCode";
 import MappedInstances from "./taskInstance/MappedInstances";
 import Logs from "./taskInstance/Logs";
 import BackToTaskSummary from "./taskInstance/BackToTaskSummary";
@@ -63,9 +64,11 @@ const tabToIndex = (tab?: string) => {
   switch (tab) {
     case "graph":
       return 1;
+    case "code":
+      return 2;
     case "logs":
     case "mapped_tasks":
-      return 2;
+      return 3;
     case "details":
     default:
       return 0;
@@ -81,6 +84,8 @@ const indexToTab = (
     case 1:
       return "graph";
     case 2:
+      return "code";
+    case 3:
       if (showMappedTasks) return "mapped_tasks";
       if (showLogs) return "logs";
       return undefined;
@@ -112,6 +117,7 @@ const Details = ({ openGroupIds, onToggleGroups, hoveredTaskState }: Props) => {
   const isGroup = !!children;
   const isGroupOrMappedTaskSummary = isGroup || isMappedTaskSummary;
   const showLogs = !!(isTaskInstance && !isGroupOrMappedTaskSummary);
+  const showDagCode = !(isTaskInstance && !isGroupOrMappedTaskSummary);
   const showMappedTasks = !!(isTaskInstance && isMappedTaskSummary && !isGroup);
 
   const [searchParams, setSearchParams] = useSearchParams();
@@ -130,7 +136,7 @@ const Details = ({ openGroupIds, onToggleGroups, hoveredTaskState }: Props) => {
   );
 
   useEffect(() => {
-    if ((!taskId || isGroup) && tabIndex > 1) {
+    if ((!taskId || isGroup) && tabIndex > 2) {
       onChangeTab(1);
     }
   }, [runId, taskId, tabIndex, isGroup, onChangeTab]);
@@ -213,6 +219,14 @@ const Details = ({ openGroupIds, onToggleGroups, hoveredTaskState }: Props) => {
               Graph
             </Text>
           </Tab>
+          {showDagCode && (
+            <Tab>
+              <MdCode size={16} />
+              <Text as="strong" ml={1}>
+                Code
+              </Text>
+            </Tab>
+          )}
           {showLogs && (
             <Tab>
               <MdReorder size={16} />
@@ -255,6 +269,11 @@ const Details = ({ openGroupIds, onToggleGroups, hoveredTaskState }: Props) => {
               hoveredTaskState={hoveredTaskState}
             />
           </TabPanel>
+          {showDagCode && (
+            <TabPanel height="100%">
+              <DagCode />
+            </TabPanel>
+          )}
           {showLogs && run && (
             <TabPanel
               pt={mapIndex !== undefined ? "0px" : undefined}