+++ airflow/providers/elasticsearch/CHANGELOG.rst
@@ -44,6 +44,9 @@ Misc
    * ``Revert "Enable individual trigger logging (#27758)" (#29472)``
    * ``Add full automation for min Airflow version for providers (#30994)``
    * ``Add mechanism to suspend providers (#30422)``
+   * ``Use '__version__' in providers not 'version' (#31393)``
+   * ``Fixing circular import error in providers caused by airflow version check (#31379)``
+   * ``Prepare docs for May 2023 wave of Providers (#31252)``
 
 4.4.0
 .....